+++ theme-contract/src/main/java/com/orange/ouds/theme/tokens/components/OudsComponentsTokens.kt
@@ -10,7 +10,7 @@
  * Software description: Android library of reusable graphical components
  */
 
-package com.orange.ouds.theme.components
+package com.orange.ouds.theme.tokens.components
 
 class OudsComponentsTokens(
     val button: OudsButtonTokens = OudsButtonTokens(),
