+++ theme-orange/src/main/java/com/orange/ouds/theme/orange/OrangeTheme.kt
@@ -13,7 +13,7 @@
 package com.orange.ouds.theme.orange
 
 import com.orange.ouds.theme.OudsCustomTheme
-import com.orange.ouds.theme.colors.OudsSemanticColors
+import com.orange.ouds.theme.tokens.semantic.OudsSemanticColors
 import kotlinx.parcelize.Parcelize
 
 const val ORANGE_THEME_NAME = "Orange"
