+++ theme-white-label/src/main/java/com/orange/ouds/theme/whitelabel/WhiteLabelSemanticColors.kt
@@ -12,8 +12,8 @@
 
 package com.orange.ouds.theme.whitelabel
 
-import com.orange.ouds.theme.colors.OudsSemanticColorValue
-import com.orange.ouds.theme.colors.OudsSemanticColors
+import com.orange.ouds.theme.tokens.semantic.OudsSemanticColorValue
+import com.orange.ouds.theme.tokens.semantic.OudsSemanticColors
 
 val WhiteLabelSemanticColors = with(WhiteLabelRawColors) {
     OudsSemanticColors(
